+++ src/attached_probe.cpp
@@ -220,6 +220,8 @@ AttachedProbe::~AttachedProbe()
     case ProbeType::uprobe:
     case ProbeType::uretprobe:
     case ProbeType::usdt:
+      if (usdt_destructor_)
+        usdt_destructor_();
       err = bpf_detach_uprobe(eventname().c_str());
       break;
     case ProbeType::tracepoint:
@@ -809,6 +811,10 @@ void AttachedProbe::attach_usdt(int pid)
       throw std::runtime_error("Error initializing context for probe: " + probe_.name);
   }
 
+  // Defer context destruction until probes are detached b/c context
+  // destruction will decrement usdt semaphore count.
+  usdt_destructor_ = [ctx]() { bcc_usdt_close(ctx); };
+
   // TODO: fn_name may need a unique suffix for each attachment on the same probe:
   std::string fn_name = "probe_" + probe_.attach_point + "_1";
 // see https://github.com/iovisor/bcc/pull/2294 for BCC_USDT_HAS_FULLY_SPECIFIED_PROBE
@@ -918,7 +924,7 @@ void AttachedProbe::attach_interval()
   int group_fd = -1;
   int cpu = 0;
 
-  uint64_t period;
+  uint64_t period = 0, freq = 0;
   if (probe_.path == "s")
   {
     period = probe_.freq * 1e9;
@@ -927,14 +933,28 @@ void AttachedProbe::attach_interval()
   {
     period = probe_.freq * 1e6;
   }
+  else if (probe_.path == "us")
+  {
+    period = probe_.freq * 1e3;
+  }
+  else if (probe_.path == "hz")
+  {
+    freq = probe_.freq;
+  }
   else
   {
     std::cerr << "invalid interval path \"" << probe_.path << "\"" << std::endl;
     abort();
   }
 
-  int perf_event_fd = bpf_attach_perf_event(progfd_, PERF_TYPE_SOFTWARE,
-      PERF_COUNT_SW_CPU_CLOCK, period, 0, pid, cpu, group_fd);
+  int perf_event_fd = bpf_attach_perf_event(progfd_,
+                                            PERF_TYPE_SOFTWARE,
+                                            PERF_COUNT_SW_CPU_CLOCK,
+                                            period,
+                                            freq,
+                                            pid,
+                                            cpu,
+                                            group_fd);
 
   if (perf_event_fd < 0)
     throw std::runtime_error("Error attaching probe: " + probe_.name);